% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia.R
\name{bed2Julia}
\alias{bed2Julia}
\title{send a bed file or df to Julia as a GenomicFeatures.IntervalCollection
with the given name}
\usage{
bed2Julia(segs = NULL, file = NULL, name = "mycollection")
}
\arguments{
\item{segs}{a dataframe with the segments.
The first columns must contain chr, start, end}

\item{file}{must be a tab separated file with
first columns must contain chr, start, end}

\item{name}{the name to assign the object to}
}
\value{
(in Julia) a GenomicFeatures.IntervalCollection without metadata
}
\description{
send a bed file or df to Julia as a GenomicFeatures.IntervalCollection
with the given name
}
\note{
Make sure GenomicFeatures is loaded in Julia
}
