% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aneuplodies.R
\name{callArmCna}
\alias{callArmCna}
\title{Wrapper around CallWcna() that calls chromosome arm aneuploidies}
\usage{
callArmCna(segs, chr, threshold = 0.95, arms = bands)
}
\arguments{
\item{chromosome}{the chromosome of interest.
must match the format of the segment table}

\item{segments}{data frame with CNA length,
chromosome and direction ("gain","loss","normal") columns}

\item{thershold}{float, the fraction of the chromosome
needed to be in a direction to call. Default = 0.95}
}
\value{
a named vector with the arm and "gain", "loss", "normal" or NA.
}
\description{
Wrapper around CallWcna() that calls chromosome arm aneuploidies
}
\note{
This assumes that there is a gap around the centromere and
that no segment can ever overlaps it.

for full documentation see callWcna.

This function estimates the length of the genome
from the length of all segments in the table
}
\seealso{
\code{\link{callWcna}}
}
