% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aneuplodies.R
\name{callWcna}
\alias{callWcna}
\title{Call Whole chromosome aneuploidies from segments based on a threshold}
\usage{
callWcna(segs, chr, threshold = 0.95)
}
\arguments{
\item{segs}{data frame with CNA length,
chromosome and direction ("gain","loss","normal") columns}

\item{chr}{the chromosome of interest. must match the format
of the segment table}

\item{thershold}{float, the fraction of the chromosome needed to be
in a direction to call. Default = 0.95}
}
\value{
"gain", "loss", "normal" or NA, a string with the status
of the chromosome
}
\description{
Call Whole chromosome aneuploidies from segments based on a threshold
}
\note{
This function estimates the length of the genome
from the length of all segments in the table

if the given chromosome is not found returns NA.

thresholds below < 0.5 always return gains as long as it passes the threshold,
regardless of whether there are more losses.
}
\seealso{
\code{\link{callArmCna}} \code{\link{fractionAltered}}
}
