% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia.R
\name{.parseC}
\alias{.parseC}
\title{private function to parse common stuff out of any GenomnePermutaiton.Jl permtest}
\usage{
.parseC(results)
}
\arguments{
\item{results}{the julia object from GenomePermutation's permtest}
}
\description{
private function to parse common stuff out of any GenomnePermutaiton.Jl permtest
}
\seealso{
\code{\link{.parseZ}}, \code{\link{.parseS}}
}
