% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aneuplodies.R
\name{fractionAltered}
\alias{fractionAltered}
\title{Calculates the fraction of the genome altrered in a given segment table}
\usage{
fractionAltered(segs)
}
\arguments{
\item{segs}{data frame with CNA length,
chromosome and direction ("gain","loss","normal") columns}
}
\value{
A named list with the fraction gained and lost
}
\description{
Calculates the fraction of the genome altrered in a given segment table
}
\note{
This function estimates the length of the genome from the length
of all segments in the table
}
