% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getName}
\alias{getName}
\title{extracts a file name from a path with leading /s and trailing extension}
\usage{
getName(path)
}
\arguments{
\item{path}{a file path}
}
\value{
a string with the file name between the last / and first \code{.}
}
\description{
extracts a file name from a path with leading /s and trailing extension
}
