% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc.R
\docType{data}
\name{hg38_centromeres}
\alias{hg38_centromeres}
\title{hg38 centromeres
cached human hg38 centromeres, using the median start/end position recorded
in}
\format{
A data frame with 24 rows and 3 variables:
\describe{
\item{chrom}{chromosome name - UCSC naming "chr1"},
\item{chromStart}{starting position - 0-indexed},
\item{chromEnd}{ending position - 0-indexed}
}
}
\source{
USCS Genome Broswer
}
\usage{
hg38_centromeres
}
\description{
hg38 centromeres
cached human hg38 centromeres, using the median start/end position recorded
in
}
\keyword{datasets}
