% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc.R
\docType{data}
\name{hg38_telomeres}
\alias{hg38_telomeres}
\title{hg38 telomeres
cached human hg38 telomeric regions using the median start/end position}
\format{
A data frame with 48 rows and 3 variables:
\describe{
\item{chrom}{chromosome name - UCSC naming "chr1"},
\item{chromStart}{starting position - 0-indexed},
\item{chromEnd}{ending position - 0-indexed}
}
}
\source{
USCS Genome Broswer
}
\usage{
hg38_telomeres
}
\description{
hg38 telomeres
cached human hg38 telomeric regions using the median start/end position
}
\keyword{datasets}
