% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia.R
\name{plotPermResults}
\alias{plotPermResults}
\title{Plots the permutation test in GenomePermutations.jl in ggplot2}
\usage{
plotPermResults(results, test)
}
\arguments{
\item{results}{the julia object from GenomePermutation's permtest}

\item{test}{the type of test, can (partially) match either "simple" or z-test"}
}
\value{
the ggplot object
}
\description{
Plots the permutation test in GenomePermutations.jl in ggplot2
}
