% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc.R
\name{queryUcsc}
\alias{queryUcsc}
\title{deal with the UCSC genome browser mySQL server
wrapper around dbGetQuery to query the UCSC genome browser mySQL server}
\usage{
queryUcsc(query, params = NULL, genome = "hg38")
}
\arguments{
\item{query}{a SELECT SQL query to run}

\item{params}{a list of parameters for the query}

\item{genome}{a UCSC genome that represents the server that will connect to}
}
\value{
the result of your query
}
\description{
deal with the UCSC genome browser mySQL server
wrapper around dbGetQuery to query the UCSC genome browser mySQL server
}
