% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{splitBp}
\alias{splitBp}
\title{Split each record in a bedfile df into 2 breakpoints}
\usage{
splitBp(bed)
}
\arguments{
\item{bed}{a bed file as a data frame. column names are unimportant
but columns 1:3 must follow bed order (chromosome, start, end)}
}
\value{
returns a bed file with
}
\description{
Split each record in a bedfile df into 2 breakpoints
}
\note{
Each record is split in two records of length 1:
Record 1 is start to start+1 and record 2 is end-1 to end.
}
