% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia.R
\name{plotPermResults}
\alias{plotPermResults}
\title{Plots the permutation test in GenomePermutations.jl in ggplot2}
\usage{
plotPermResults(results)
}
\arguments{
\item{results}{the Results from the permutation test in Julia}
}
\value{
the ggplot object
}
\description{
Plots the permutation test in GenomePermutations.jl in ggplot2
}
